refactor(gencom): update code and docs
Performed wide-ranging refactoring to the `gencom` codebase, including
more detailed and improved comments and adjustments to string
representation and struct implementation. Expanded the `GitInterface` to
the `GitClient`, and the usage and application of this new interface in
functions such as `NewGitCommand`, `GetDiff`, and new tests. Adjusted
commit rendering and user flow.